# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WebhookPayloadAction =
  (typeof WebhookPayloadAction)[keyof typeof WebhookPayloadAction];

export const WebhookPayloadAction = {
  BUY: "BUY",
  SELL: "SELL",
} as const;

export interface WebhookPayload {
  /** Trading symbol e.g. BTCUSD */
  symbol: string;
  action: WebhookPayloadAction;
  /** Signal price */
  price?: number;
}

export interface WebhookResponse {
  received: boolean;
  signalId: number;
}

export type SignalAction = (typeof SignalAction)[keyof typeof SignalAction];

export const SignalAction = {
  BUY: "BUY",
  SELL: "SELL",
} as const;

export interface Signal {
  id: number;
  symbol: string;
  action: SignalAction;
  price?: number | null;
  processed: boolean;
  createdAt: string;
}

export interface SignalListResponse {
  signals: Signal[];
  total: number;
  limit: number;
  offset: number;
}

export interface LatestSignalResponse {
  signal: Signal | null;
}

export interface Symbol {
  id: number;
  name: string;
  xmUrl: string;
  createdAt: string;
}

export interface CreateSymbolRequest {
  name: string;
  xmUrl: string;
}

export interface DeleteResponse {
  deleted: boolean;
}

export interface Settings {
  id: number;
  /** Take profit in USD */
  tpAmount: number;
  /** Stop loss in USD */
  slAmount: number;
  automationEnabled: boolean;
}

export interface UpdateSettingsRequest {
  tpAmount?: number;
  slAmount?: number;
  automationEnabled?: boolean;
}

export type ListSignalsParams = {
  limit?: number;
  offset?: number;
  processed?: boolean;
};