# lib/api-zod/src/generated/types/signal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignalAction } from "./signalAction";

export interface Signal {
  id: number;
  symbol: string;
  action: SignalAction;
  price?: number | null;
  processed: boolean;
  createdAt: Date;
}