# lib/api-zod/src/generated/types/signalListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Signal } from "./signal";

export interface SignalListResponse {
  signals: Signal[];
  total: number;
  limit: number;
  offset: number;
}