# lib/api-zod/src/generated/types/webhookPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WebhookPayloadAction } from "./webhookPayloadAction";

export interface WebhookPayload {
  /** Trading symbol e.g. BTCUSD */
  symbol: string;
  action: WebhookPayloadAction;
  /** Signal price */
  price?: number;
}