# lib/db/src/schema/settings.ts
import { pgTable, serial, real, boolean } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const settingsTable = pgTable("settings", {
  id: serial("id").primaryKey(),
  tpAmount: real("tp_amount").notNull().default(2.0),
  slAmount: real("sl_amount").notNull().default(2.0),
  automationEnabled: boolean("automation_enabled").notNull().default(true),
});

export const insertSettingsSchema = createInsertSchema(settingsTable).omit({ id: true });
export type InsertSettings = z.infer<typeof insertSettingsSchema>;
export type Settings = typeof settingsTable.$inferSelect;